refactor(functions): migrate to firebase functions v2 api
- merge `indexOnWriteSecondGen` into `indexOnWrite` using v2 api
- merge `backfillSecondGen` into `backfill` using v2 api
- update event trigger implementation in `extension.yaml`
- remove V2 API logging context
