+++ extension.yaml
@@ -19,34 +19,6 @@ billingRequired: true
 resources:
   - name: indexOnWrite
     description: A function that indexes data into Typesense when it's triggered by Firestore changes
-    type: firebaseextensions.v1beta.function
-    properties:
-      runtime: nodejs18
-      sourceDirectory: functions
-      timeout: 540s
-      location: ${LOCATION}
-      eventTrigger:
-        eventType: providers/cloud.firestore/eventTypes/document.write
-        resource: projects/${param:PROJECT_ID}/databases/(default)/documents/${param:FIRESTORE_COLLECTION_PATH}/{documentID}
-
-  - name: backfill
-    description: >-
-      A function that backfills data from a Firestore collection into Typesense,
-      triggered when a Firestore document with the path `typesense_sync/backfill` has the contents of `trigger: true`.
-    type: firebaseextensions.v1beta.function
-    properties:
-      runtime: nodejs18
-      sourceDirectory: functions
-      timeout: 540s
-      maxInstances: 1
-      availableMemoryMb: 2048
-      location: ${LOCATION}
-      eventTrigger:
-        eventType: providers/cloud.firestore/eventTypes/document.write
-        resource: projects/${param:PROJECT_ID}/databases/(default)/documents/typesense_sync/backfill
-
-  - name: indexOnWriteSecondGen
-    description: A function that indexes data into Typesense when it's triggered by Firestore changes
     type: firebaseextensions.v1beta.v2.function
     properties:
       sourceDirectory: functions
@@ -64,7 +36,7 @@ resources:
             value: ${FIRESTORE_COLLECTION_PATH}/{documentID}
             operator: match-path-pattern
 
-  - name: backfillSecondGen
+  - name: backfill
     description: >-
       A function that backfills data from a Firestore collection into Typesense,
       triggered when a Firestore document with the path `typesense_sync/backfill` has the contents of `trigger: true`.
