+++ functions/src/indexOnWrite.js
@@ -1,35 +1,29 @@
-const functions = require("firebase-functions/v1");
+const {debug} = require("firebase-functions/logger");
 const config = require("./config.js");
-const createTypesenseClient = require("./createTypesenseClient.js");
 const utils = require("./utils.js");
+const createTypesenseClient = require("./createTypesenseClient.js");
+const {onDocumentWritten} = require("firebase-functions/v2/firestore");
 
-module.exports = functions.firestore.document(config.firestoreCollectionPath)
-    .onWrite(async (snapshot, context) => {
-      const typesense = createTypesenseClient();
+exports.indexOnWrite = onDocumentWritten(`${config.firestoreCollectionPath}/{docId}`, async (snapshot, _) => {
+  const typesense = createTypesenseClient();
 
-      if (snapshot.after.data() == null) {
-        // Delete
-        const documentId = snapshot.before.id;
-        functions.logger.debug(`[V1 API]: Deleting document ${documentId}`);
-        return await typesense
-            .collections(encodeURIComponent(config.typesenseCollectionName))
-            .documents(encodeURIComponent(documentId))
-            .delete();
-      } else {
-        // Create / update
+  if (snapshot.data.after.data() == null) {
+    // Delete
+    const documentId = snapshot.data.before.id;
+    debug(`Deleting document ${documentId}`);
+    return await typesense.collections(encodeURIComponent(config.typesenseCollectionName)).documents(encodeURIComponent(documentId)).delete();
+  } else {
+    // Create / update
 
-        // snapshot.after.ref.get() will refetch the latest version of the document
-        const latestSnapshot = await snapshot.after.ref.get();
-        const typesenseDocument = await utils.typesenseDocumentFromSnapshot(latestSnapshot, context.params);
+    // snapshot.after.ref.get() will refetch the latest version of the document
+    const latestSnapshot = await snapshot.data.after.ref.get();
+    const typesenseDocument = await utils.typesenseDocumentFromSnapshot(latestSnapshot, snapshot.params);
 
-        if (config.shouldLogTypesenseInserts) {
-          functions.logger.debug(`[V1 API]: Upserting document ${JSON.stringify(typesenseDocument)}`);
-        } else {
-          functions.logger.debug(`[V1 API]: Upserting document ${typesenseDocument.id}`);
-        }
-        return await typesense
-            .collections(encodeURIComponent(config.typesenseCollectionName))
-            .documents()
-            .upsert(typesenseDocument);
-      }
-    });
+    if (config.shouldLogTypesenseInserts) {
+      debug(`Upserting document ${JSON.stringify(typesenseDocument)}`);
+    } else {
+      debug(`Upserting document ${typesenseDocument.id}`);
+    }
+    return await typesense.collections(encodeURIComponent(config.typesenseCollectionName)).documents().upsert(typesenseDocument);
+  }
+});
